+++ app/src/main/java/org/simple/clinic/searchresultsview/PatientSearchUi.kt
@@ -1,7 +1,0 @@
-package org.simple.clinic.searchresultsview
-
-interface PatientSearchUi {
-  fun updateSearchResults(results: PatientSearchResults)
-  fun searchResultClicked(searchResultClickedEvent: SearchResultClicked)
-  fun registerNewPatient(registerNewPatientEvent: RegisterNewPatient)
-}