+++ app/src/main/java/org/simple/clinic/searchresultsview/PatientSearchViewController.kt
@@ -4,9 +4,7 @@ import io.reactivex.Observable
 import io.reactivex.ObservableSource
 import io.reactivex.ObservableTransformer
 import io.reactivex.rxkotlin.ofType
-import io.reactivex.rxkotlin.withLatestFrom
 import org.simple.clinic.ReplayUntilScreenIsDestroyed
-import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.bp.BloodPressureMeasurement
 import org.simple.clinic.facility.FacilityRepository
 import org.simple.clinic.patient.PatientRepository
@@ -14,7 +12,7 @@ import org.simple.clinic.user.UserSession
 import org.simple.clinic.widgets.UiEvent
 import javax.inject.Inject
 
-typealias Ui = PatientSearchUi
+typealias Ui = SearchResultsUi
 typealias UiChange = (Ui) -> Unit
 
 class PatientSearchViewController @Inject constructor(
@@ -26,14 +24,9 @@ class PatientSearchViewController @Inject constructor(
 
   override fun apply(events: Observable<UiEvent>): ObservableSource<UiChange> {
     val replayedEvents = ReplayUntilScreenIsDestroyed(events)
-        .compose(ReportAnalyticsEvents())
         .replay()
 
-    return Observable.mergeArray(
-        openPatientSummary(replayedEvents),
-        createNewPatient(replayedEvents),
-        populateSearchResults(replayedEvents)
-    )
+    return populateSearchResults(replayedEvents)
   }
 
   private fun populateSearchResults(events: Observable<UiEvent>): Observable<UiChange> {
@@ -54,23 +47,6 @@ class PatientSearchViewController @Inject constructor(
           { ui: Ui ->
             ui.updateSearchResults(results)
           }
-        }
-  }
-
-  private fun openPatientSummary(events: Observable<UiEvent>): Observable<UiChange> {
-    return events.ofType<SearchResultClicked>()
-        .map { { ui: Ui -> ui.searchResultClicked(it) } }
-  }
-
-  private fun createNewPatient(events: Observable<UiEvent>): Observable<UiChange> {
-    val searchPatientWithCriteriaStream = events
-        .ofType<SearchPatientWithCriteria>()
-        .map { it.criteria }
-
-    return events.ofType<RegisterNewPatientClicked>()
-        .withLatestFrom(searchPatientWithCriteriaStream)
-        .map { (_, patientSearchCriteria) ->
-          { ui: Ui -> ui.registerNewPatient(RegisterNewPatient(patientSearchCriteria)) }
         }
   }
 }