+++ app/src/main/java/org/simple/clinic/searchresultsview/SearchResultsEffect.kt
@@ -1,0 +1,3 @@
+package org.simple.clinic.searchresultsview
+
+sealed class SearchResultsEffect