+++ app/src/main/java/org/simple/clinic/searchresultsview/SearchResultsUi.kt
@@ -1,0 +1,5 @@
+package org.simple.clinic.searchresultsview
+
+interface SearchResultsUi {
+  fun updateSearchResults(results: PatientSearchResults)
+}