+++ app/src/main/java/org/simple/clinic/searchresultsview/SearchResultsViewEvents.kt
@@ -4,22 +4,9 @@ import org.simple.clinic.patient.PatientSearchCriteria
 import org.simple.clinic.patient.PatientSearchCriteria.Name
 import org.simple.clinic.patient.PatientSearchCriteria.PhoneNumber
 import org.simple.clinic.widgets.UiEvent
-import java.util.UUID
 
 object SearchResultsViewCreated : UiEvent {
   override val analyticsName = "Search Results:View Created"
-}
-
-data class SearchResultClicked(val patientUuid: UUID) : UiEvent {
-  override val analyticsName = "Search Results:Search Result Clicked"
-}
-
-object RegisterNewPatientClicked : UiEvent {
-  override val analyticsName = "Search Results:Register New Patient Clicked"
-}
-
-data class RegisterNewPatient(val criteria: PatientSearchCriteria) : UiEvent {
-  override val analyticsName = "Search Results:Register New Patient"
 }
 
 data class SearchPatientWithCriteria(val criteria: PatientSearchCriteria) : UiEvent {